Changed offset to a double to fix a crash when the offset is not an int Don't show locality if it is null

+++ app/src/main/java/me/danbeneventano/weather/dataclasses/WeatherResponse.kt
@@ -1,7 +1,7 @@
 package me.danbeneventano.weather.dataclasses
 
 data class WeatherResponse(
-        val offset: Int? = null,
+        val offset: Double? = null,
         val currently: DataPoint,
         val timezone: String,
         val latitude: Double,
